save publication via channel, #165
