+++ test/controllers/consent_controller_test.exs
@@ -14,7 +14,7 @@ defmodule Bep.ConsentControllerTest do
     assert html_response(conn, 200) =~ "I agree"
   end
 
-  @tag login_as: %{email: "email@example.com"}
+  @tag login_as: %{email: "email@example.com", id: 1}
   test "GET / redirect to /search when logged in", %{conn: conn} do
     conn = get conn, "/consent"
     assert html_response(conn, 302)
