+++ web/controllers/auth.ex
@@ -6,6 +6,7 @@ defmodule Bep.Auth do
   import Phoenix.Controller
   import Comeonin.Bcrypt, only: [checkpw: 2, dummy_checkpw: 0]
   alias Bep.Router.Helpers
+  alias Phoenix.Token
 
   def init(opts) do
     Keyword.fetch!(opts, :repo)
@@ -15,9 +16,9 @@ defmodule Bep.Auth do
     user_id = get_session(conn, :user_id)
     cond do
       user = conn.assigns[:current_user] ->
-        conn
+        put_current_user(conn, user)
       user = user_id && repo.get(Bep.User, user_id) ->
-        assign(conn, :current_user, user)
+        put_current_user(conn, user)
       true ->
         assign(conn, :current_user, nil)
     end
@@ -37,9 +38,17 @@ defmodule Bep.Auth do
 
   def login(conn, user) do
     conn
-    |> assign(:current_user, user)
+    |> put_current_user(user)
     |> put_session(:user_id, user.id)
     |> configure_session(renew: true)
+  end
+
+  defp put_current_user(conn, user) do
+    token = Token.sign(conn, "user socket", user.id)
+
+    conn
+    |> assign(:current_user, user)
+    |> assign(:user_token, token)
   end
 
   def login_by_email_and_pass(conn, email, given_pass, opts) do
