+++ web/templates/layout/app.html.eex
@@ -30,6 +30,7 @@
     <link rel="mask-icon" href="images/favicon/safari-pinned-tab.svg" color="#5bbad5">
     <meta name="theme-color" content="#ffffff">
     <title>Best Evidence</title>
+    <script>window.userToken = "<%= assigns[:user_token] %>"</script>
     <link rel="stylesheet" href="<%= static_path(@conn, "/css/app.css") %>">
   </head>
